Add new dependencies to Cargo.lock
This commit introduces several new dependencies such as `android-tzdata`, `android_system_properties`, `byteorder`, and `chrono` to `Cargo.lock`. These additions are brought in through updates or inclusion of other packages like `iana-time-zone`, `lzma-sys`, and `xattr`. The changes ensure required packages and their dependencies are tracked and included for the project.